Moved VideoGrid to separate file
